% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{acf_arma}
\alias{acf_arma}
\title{Autocorrelation function estimation for casual ARMA(p,q) process}
\usage{
acf_arma(phi, theta, lag = 20, dig = 5)
}
\arguments{
\item{phi}{A vector, containing coefficient of AR part as array. it require to have 1 as leading coefficient.}

\item{theta}{A vector, containing coefficient of MA part as array. it require to have 1 as leading coefficient.}

\item{lag}{the desired maximun lag , The default value is 20.}

\item{dig}{the desired rounding digit of return value, the default value is 5}
}
\value{
None
}
\description{
The function acf_arma() computes estimates of the autocorrelation/autocoveriance up to given lag.
}
\details{
It retruns a data frame contain lag (lags), autocovariance (gamma) and autocorrelation (rho).
}
\examples{
# Define time series data in a vector, x
phi <- c(1, -1, 1 / 4)
theta <- c(1, 1)

# Use acf_arma function
acf_arma(phi, theta)
acf_arma(phi, theta, 20, 8)

# Another example
phi <- c(1, rnorm(100))
theta <- c(1,rnorm(50))

acf_arma(phi, theta)
acf_arma(phi, theta, 10, 8)

}
\references{
Brockwell, P. J., & Davis, R. A. (2013). Time series: theory and methods. New York, NY: Springer.
}
\author{
Ankai Liu, \email{ankailiu@trentu.ca}
}
\keyword{ARMA}
\keyword{acf,}
\keyword{autocorrelation,}
\keyword{autocovariance,}
