% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\name{as.character}
\alias{as.character}
\title{Convert task object to character, can be dumped to file and used to instantiate future Task objects}
\usage{
\S4method{as.character}{Task}(x, ...)
}
\description{
Convert task object to character, can be dumped to file and used to instantiate future Task objects
}
\author{
Marcin Imielinski
}
