% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{numeric number of trials}

\item{prob}{probability of success}
}
\value{
data frame with number of successes, probabiility, and cumulative probability
}
\description{
Calculates likelihood of each bin in binomial distribution and cumulative probability
}
\examples{
# binomial cumulative distribution
bin_cumulative(trials = 5, prob = 0.5)
}
