% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{numeric number of trials}

\item{prob}{probability of success}
}
\value{
data frame with number of successes and probability
}
\description{
Calculates likelihood of each bin in binomial distribution
}
\examples{
# binomial probability distribution
bin_distribution(trials = 5, prob = .5)
}
