% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_functions.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{numeric number of trials}

\item{prob}{probability of success}
}
\value{
List of class binvar with trials and prob
}
\description{
Calculates likelihood of each bin in binomial distribution and cumulative probability
}
\examples{
bin_variable(5, .5)
}
