% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blank}
\alias{blank}
\title{Parameter table created}
\format{
A 2 column 1 observation
}
\source{
created on September 5, 2020
}
\usage{
blank
}
\description{
Parameter table created
}
\keyword{datasets}
