% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3a_trialdata_analysis_input_functions.R
\name{check_treatment_arm}
\alias{check_treatment_arm}
\title{Function to return treatment arm}
\usage{
check_treatment_arm(arm)
}
\arguments{
\item{arm}{the arm of the trial}
}
\value{
0, if success -1, if failure
}
\description{
Function to return treatment arm
}
\examples{
check_treatment_arm("control")
}
