% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_model_functions.R
\name{combine_state}
\alias{combine_state}
\title{Join health states}
\usage{
combine_state(...)
}
\arguments{
\item{...}{any additional objects}
}
\value{
joined health states
}
\description{
Join health states
}
\details{
checking each state is a health state and join them
}
\examples{
a <- health_state("IT", 100, 0.4, 0, FALSE)
b <- health_state("PT", 100, 0.4, 0, FALSE)
combine_state(a, b)
}
