% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_cost_analysis_functions.R
\name{convert_volume_basis}
\alias{convert_volume_basis}
\title{Convert volume  to given basis}
\usage{
convert_volume_basis(given_unit, basis = "ml")
}
\arguments{
\item{given_unit}{given unit}

\item{basis}{given basis, default is "ml"}
}
\value{
converted unit
}
\description{
Convert volume  to given basis
}
\examples{
convert_volume_basis("ml", "liter")
}
