% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dedup}
\alias{dedup}
\title{dedup}
\usage{
dedup(x, suffix = ".")
}
\arguments{
\item{x}{vector of files}

\item{suffix}{string suffix (default = '.')}
}
\description{
relabels duplicates in a character vector with .1, .2, .3
(where "." can be replaced by any user specified suffix)
}
\author{
Marcin Imielinski
}
