% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\name{err}
\alias{err}
\title{returns vector of paths to stderr files associated with this job object}
\usage{
\S4method{err}{Job}(.Object)
}
\description{
returns vector of paths to stderr files associated with this job object
}
\author{
Marcin Imielinski
}
