% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_notused_functions.R
\name{find_param_from_def_prob_distrbn}
\alias{find_param_from_def_prob_distrbn}
\title{Function to return parameters from a given expression separated by comma,}
\usage{
find_param_from_def_prob_distrbn(expr)
}
\arguments{
\item{expr}{an expression}
}
\value{
parameters in the expression expr
}
\description{
Function to return parameters from a given expression separated by comma,
}
\details{
It will return the parameters of the distribution separated by commas
and given in usual notation as brackets
It will identify those in between first occurrence of "( "and last occurrence of ")"
and from the characters in between search for comma to indicate different parameters
then it will extract (from those extracted parameters separated by commas)
that on the left side of "equal" sign
find_param_from_def_prob_distrbn("gamma(mean = sqrt(2), b =17)") will be ok
but find_param_from_def_prob_distrbn("gamma(shape, scale")) and
find_param_from_def_prob_distrbn("gamma(shape =1 & scale =1")) will throw error
}
\examples{
find_param_from_def_prob_distrbn("gamma(mean = 10, sd =1)")
}
