% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\docType{class}
\name{Flow-class}
\alias{Flow-class}
\title{Class representing a collection of tasks}
\description{
S4 class for \code{Flow}

A Flow is just a set of tasks / jobs applied to a set of entities.  This class is currently only
for monitoring progress / visualization of completeness, i.e. via a pie-network. 

The taskdir is a directory which has .task objects, which are scraped via Task instantiator.
}
\section{Slots}{

\describe{
  \item{nodes}{data.table of nodes which are either annotations or Task objects}
  \item{edges}{data.table of edges which are connect (input) annotations to Task objects or Task objects to output annotations}
  \item{entities}{keyed data.table of entities to which tasks are (optionally) applied to} 
}
}

\author{
Marcin Imielinski
}
