% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cpp.R
\name{generate_output_interface}
\alias{generate_output_interface}
\title{Generate C++ for passing output from model fit into R}
\usage{
generate_output_interface(dest)
}
\arguments{
\item{dest}{The destination to write generated code to.}
}
\value{
Nothing, called to generate code in src dir
}
\description{
This generates using metadata from \code{src/model_output.yml}
}
\keyword{internal}
