% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOutputTableNames.R
\name{getOutputTableNames}
\alias{getOutputTableNames}
\title{A getOutputTableNames Function}
\usage{
getOutputTableNames(iProcess)
}
\description{
Function getOutputTableNames.
}
\examples{
getOutputTableNames()
}
