% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\name{ids}
\alias{ids}
\title{returns vector of entity ids associated with this Job objedct}
\usage{
\S4method{ids}{Job}(.Object)

\S4method{ids}{Job}(.Object)
}
\description{
returns vector of entity ids associated with this Job objedct

returns vector of entity ids associated with this Job objedct
}
\author{
Marcin Imielinskix

Marcin Imielinskix
}
