% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\name{io_c<-}
\alias{io_c<-}
\title{Sets ionice parameters associated with Job object}
\usage{
\S4method{io_c}{Job}(.Object) <- value
}
\description{
setting io_c parameters associated with job object
}
\author{
Kevin Hadi
}
