% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\docType{class}
\name{Module-class}
\alias{Module-class}
\title{Module class}
\description{
Class \code{Module} defines an unconfigured module that is scraped from a "module" or "hydrant.deploy" file.
}
\section{Slots}{

\describe{
  \item{name}{character name of module (by default the name of the module directory)}
  \item{sourcedir}{character path of firehose module directory containing code and hydrant.deploy file}
  \item{args}{vector of argument names that need to be populated for this module to be configured into a task}
  \item{cmd}{formatted string to be populated by args at time of instantiation of this module into Job}
  \item{stamp}{version stamp of this module}
}
}

\author{
Marcin Imielinski
}
