% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\name{out}
\alias{out}
\title{returns vector of paths to stdout files associated with this job object}
\usage{
\S4method{out}{Job}(.Object)
}
\description{
returns vector of paths to stdout files associated with this job object
}
\author{
Marcin Imielinski
}
