% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\name{outdir}
\alias{outdir}
\title{Gets vector of output directories associated with the jobs in this Job object}
\usage{
\S4method{outdir}{Job}(.Object)
}
\description{
Gets vector of output directories associated with the jobs in this Job object
}
\author{
Marcin Imielinski
}
