% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\name{plot.Flow}
\alias{plot.Flow}
\title{plot.Flow}
\usage{
\method{plot}{Flow}(x, y, paths.only = FALSE, seed = 42, layout = layout_with_fr)
}
\arguments{
\item{x}{Flow object}
}
\description{
Plots a Flow as a graph of pie charts
}
\author{
Marcin Imielinski
}
