% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3d_decision_analysis_functions.R
\name{plot_efficiency_frontier}
\alias{plot_efficiency_frontier}
\title{Plot efficiency frontier}
\usage{
plot_efficiency_frontier(results_cea, threshold)
}
\arguments{
\item{results_cea}{results from cea (from calculate_icer_nmb method)}

\item{threshold}{threshold value}
}
\value{
plot
}
\description{
Plot efficiency frontier
}
