% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotglyphs.r
\name{plotGlyphs}
\alias{plotGlyphs}
\title{Visualise Spatial Uncertainty using Glyphs.}
\usage{
plotGlyphs(simulations, density, title = NULL, xlab = NULL, ylab = NULL,
  grid = T, axis = T, mean_col = NULL, glyph_col = NULL,
  legend = NULL)
}
\arguments{
\item{simulations}{Object of class simulations containing continuous data}

\item{density}{Numeric. Gives spacing of density of glyphs.}

\item{title, xlab, ylab}{Character vector giving plot title, x axis label and y
axis label respectively.}

\item{grid}{Logical. Add grid to background of graph}

\item{axis}{Logical. Add axis ticks and labels to graph}

\item{mean_col}{\code{\link[grDevices]{colorRampPalette}}. Colours to
be used for plotting the mean values of the realisations.}

\item{glyph_col}{Character giving the colour of the glyphs representing the
standard deviation of the realisations.}

\item{legend}{Character giving the legend titles. To be provided in format
c("legend1", "legend2").}
}
\value{
Nothing at all
}
\description{
Allows for visualising spatial uncertainty for continuous data using glyphs.
}
\examples{
YOLO
}

