% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\name{qprior}
\alias{qprior}
\title{Gets qprior value}
\usage{
\S4method{qprior}{Job}(.Object)
}
\description{
sets qsub -p parameter
this is the parameter that specifies priority
can be value between -1023 to 2014
default is 0
}
\author{
Kevin Hadi
}
