% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBioticDataFromXml.R
\name{readBioticDataFromXml}
\alias{readBioticDataFromXml}
\title{A readBioticDataFromXml Function}
\usage{
readBioticDataFromXml()
}
\arguments{
\item{no}{parameters.}
}
\description{
This function is used to read Biotic Data From Xml file.
}
\examples{
readBioticDataFromXml()
}
