% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_model_functions.R
\name{set_var_state}
\alias{set_var_state}
\title{Set the attribute for the health state}
\usage{
set_var_state(state, var, new_value)
}
\arguments{
\item{state}{object of class health state}

\item{var}{attribute of the health state}

\item{new_value}{new value to be assigned}
}
\value{
modified health state
}
\description{
Set the attribute for the health state
}
\details{
After checking the given state is a health state
the value of the variable is set
if the value is not numeric, it is being parsed to form an expression
}
\examples{
set_var_state(health_state("IT", 100, 0.4, 0, FALSE), "cost", 1)
}
