% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\name{skill}
\alias{skill}
\title{Kills any running Slurm jobs associated with this Job object}
\usage{
\S4method{skill}{Job}(.Object, jid = NULL)
}
\description{
Kills any running Slurm jobs associated with this Job object
}
\author{
Kevin Hadi
}
