% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flow.R
\name{tasks}
\alias{tasks}
\title{returns original keyed data.table of entities associated with this Flow object}
\usage{
\S4method{tasks}{Flow}(.Object)
}
\description{
returns original keyed data.table of entities associated with this Flow object
}
\author{
Marcin Imielinski
}
