% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_geojson_points.R
\name{test_geojson_points}
\alias{test_geojson_points}
\title{A test_geojson_points Function}
\usage{
test_geojson_points()
}
\arguments{
\item{no}{parameters.}
}
\description{
This function to test geojson points.
}
\examples{
test_geojson_points()
}
\keyword{geojson}
