% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_geojson_polygon.R
\name{test_geojson_polygon}
\alias{test_geojson_polygon}
\title{A test_geojson_polygon Function}
\usage{
test_geojson_polygon()
}
\arguments{
\item{no}{parameters.}
}
\description{
This function to test geojson polygon.
}
\examples{
test_geojson_polygon()
}
\keyword{geojson}
